Change state/county load script into a migration
Based on this documentation:
https://docs.djangoproject.com/en/2.1/howto/initial-data/
https://docs.djangoproject.com/en/2.1/topics/migrations/#data-migrations

Using a data migration means the initial state and county instances will
be loaded automatically when the database is set up, without having to
run a separate command. This includes for unit tests - Django sets up a
new blank database when running tests, and will run the data migration
automatically, meaning the states and counties will be available when
unit testing.