Add models to store percentile-values
To chart the national percentile curve for a data set, we need to
preserve the calculated values for each of the percentiles, e.g. if we
calculate that 45% of the values in the data set are <= the value 650,
then we need to store the pair (0.45, 650). If we store a pair like this
for each of the 1000 percentiles we calculate, we should be able to plot
a smooth curve without storing redundant information on data point
instances, or without needing to send all 3144 data points to the client.