Split creating percentiles into two steps, and store percentiles
One function is from points to percentile values,
Another function takes a list of points and a list of percentile-values
and assigns a percentile to each point. (Previously a single function
did this.)

This allows us to create models for the percentile values and save them
as part of creating a data set.