Cherry-pick of 6f43ba
Changes from Feyi. Original messages:

I tried linking the 'landing' page to the home icon for demo, but it wasn't working. It gave me an error. I don't want to mess with the urls, so I left it at that. 

Removed the 'remember me' on the log in page.

Fixes DES-125
Closes DES-123