Add computed field to county objects
Uses a custom model manager (US_County_Manager) to add a computed field
to US_County instances as they are queried. Doing it this way (rather
than adding a regular python @property or method to the model class)
allows the computed property to be used inside QuerySets (e.g., when
filtering).

Source: https://stackoverflow.com/a/42491803/5111071