Add a database fixture containing populated indicators
Created a database fixture file containing 9 health indicators and 7
data sets carved from the County Health Rankings 2018 report.

To use this file, starting from a acompletely blank database:
- python manage.py migrate
- python manage.py createsuperuser
- python manage.py loaddata initial_indicators

You will then have a database with the states and counties (from running
the migrations), an admin user (from createsuperuser), and a number of
full data sets (from the loaddata command).

The data sets / data points in the fixture include foreign keys to a user,
counties, and uploaded files. After applying the fixture *these may not
match as expected* because:
- users and counties have auto-incremented primary keys
- there are no uploaded documents

As such it would not be best practice to load the production database
from this fixture, but it is possible.