User Management
Two forms classes add fields to Django's create user template and add a new field for user type (extension of user model). Migration for new model to extend user model by one field. Base file has a new option in the dropdown for user management. New template for creating a new user. New template for user management, simply contains links to 3 management options specified in Jira issues. Two URLs added for the 2 templates. Added to privileged views: a function  that is a custom decorator to verify if user is admin or assistant, a view for the user management template, and a view for the create user template.